# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/web/scrapper": {
      "post": {
        "tags": [
          "Website Scrapper"
        ],
        "summary": "Website Data Extraction.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "access data",
            "in": "body",
            "description": "JSON data for the api",
            "required": true,
            "schema": {
              "$ref": "#/definitions/webScrapperRequestSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/webScrapperResponseSchema"
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    },
    "/api/web/scrapper/profiles": {
      "post": {
        "tags": [
          "Website Scrapper"
        ],
        "summary": "Web Profile Information Extraction.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "access data",
            "in": "body",
            "description": "JSON data for the api",
            "required": true,
            "schema": {
              "$ref": "#/definitions/webScrapperRequestSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/webScrapperResponseSchema"
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    },
    "/api/pdf/scrap": {
      "post": {
        "tags": [
          "PDF Scrapper"
        ],
        "summary": "PDF scrap.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "access data",
            "in": "body",
            "description": "JSON data for the api",
            "required": true,
            "schema": {
              "$ref": "#/definitions/webScrapperRequestSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/webScrapperResponseSchema"
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    }
  },
  "definitions": {
    "webScrapperRequestSchema": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        }
      }
    },
    "webScrapperResponseSchema": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        }
      }
    },
    "ValidateResponse": {
      "type": "object",
      "properties": {
        "green": {
          "type": "string"
        },
        "red": {
          "type": "string"
        }
      }
    }
  }
}